+++ rust/dbn/src/decode/dbn/async.rs
@@ -665,15 +665,15 @@ mod tests {
             DbnEncodable,
         },
         rtype, CbboMsg, Error, ErrorMsg, ImbalanceMsg, InstrumentDefMsg, MboMsg, Mbp10Msg, Mbp1Msg,
-        OhlcvMsg, RecordHeader, Result, Schema, StatMsg, TbboMsg, TradeMsg, WithTsOut,
+        OhlcvMsg, RecordHeader, Result, Schema, StatMsg, StatusMsg, TbboMsg, TradeMsg, WithTsOut,
     };
 
     #[rstest]
     #[case::mbo(Schema::Mbo, MboMsg::default())]
     #[case::trades(Schema::Trades, TradeMsg::default())]
     #[case::tbbo(Schema::Cbbo, CbboMsg::default_for_schema(Schema::Cbbo))]
-    #[case::tbbo(Schema::Tbbo, TbboMsg::default_for_schema(Schema::Tbbo))]
-    #[case::mbp1(Schema::Mbp1, Mbp1Msg::default_for_schema(Schema::Mbp1))]
+    #[case::tbbo(Schema::Tbbo, TbboMsg::default())]
+    #[case::mbp1(Schema::Mbp1, Mbp1Msg::default())]
     #[case::mbp10(Schema::Mbp10, Mbp10Msg::default())]
     #[case::ohlcv1d(Schema::Ohlcv1D, OhlcvMsg::default_for_schema(Schema::Ohlcv1D))]
     #[case::ohlcv1h(Schema::Ohlcv1H, OhlcvMsg::default_for_schema(Schema::Ohlcv1H))]
@@ -682,6 +682,7 @@ mod tests {
     #[case::definitions(Schema::Definition, InstrumentDefMsg::default())]
     #[case::imbalance(Schema::Imbalance, ImbalanceMsg::default())]
     #[case::statistics(Schema::Statistics, StatMsg::default())]
+    #[case::status(Schema::Status, StatusMsg::default())]
     #[tokio::test]
     async fn test_dbn_identity<R: DbnEncodable + HasRType + PartialEq + Clone>(
         #[case] schema: Schema,
@@ -716,8 +717,8 @@ mod tests {
     #[case::mbo(Schema::Mbo, MboMsg::default())]
     #[case::trades(Schema::Trades, TradeMsg::default())]
     #[case::cbbo(Schema::Cbbo, CbboMsg::default_for_schema(Schema::Cbbo))]
-    #[case::tbbo(Schema::Tbbo, TbboMsg::default_for_schema(Schema::Tbbo))]
-    #[case::mbp1(Schema::Mbp1, Mbp1Msg::default_for_schema(Schema::Mbp1))]
+    #[case::tbbo(Schema::Tbbo, TbboMsg::default())]
+    #[case::mbp1(Schema::Mbp1, Mbp1Msg::default())]
     #[case::mbp10(Schema::Mbp10, Mbp10Msg::default())]
     #[case::ohlcv1d(Schema::Ohlcv1D, OhlcvMsg::default_for_schema(Schema::Ohlcv1D))]
     #[case::ohlcv1h(Schema::Ohlcv1H, OhlcvMsg::default_for_schema(Schema::Ohlcv1H))]
@@ -726,6 +727,7 @@ mod tests {
     #[case::definitions(Schema::Definition, InstrumentDefMsg::default())]
     #[case::imbalance(Schema::Imbalance, ImbalanceMsg::default())]
     #[case::statistics(Schema::Statistics, StatMsg::default())]
+    #[case::status(Schema::Status, StatusMsg::default())]
     #[tokio::test]
     async fn test_dbn_zstd_identity<R: DbnEncodable + HasRType + PartialEq + Clone>(
         #[case] schema: Schema,