+++ rust/dbn/src/decode/dbn/sync.rs
@@ -685,8 +685,8 @@ mod tests {
             dbn::Encoder, DbnEncodable, DbnRecordEncoder, DynWriter, EncodeDbn, EncodeRecord,
         },
         rtype, CbboMsg, Compression, Error, ErrorMsg, ImbalanceMsg, InstrumentDefMsg, MboMsg,
-        Mbp10Msg, Mbp1Msg, MetadataBuilder, OhlcvMsg, RecordHeader, Result, StatMsg, TbboMsg,
-        TradeMsg, WithTsOut, SYMBOL_CSTR_LEN,
+        Mbp10Msg, Mbp1Msg, MetadataBuilder, OhlcvMsg, RecordHeader, Result, StatMsg, StatusMsg,
+        TbboMsg, TradeMsg, WithTsOut, SYMBOL_CSTR_LEN,
     };
 
     #[test]
@@ -736,18 +736,8 @@ mod tests {
     #[rstest]
     #[case::uncompressed_mbo_v1(1, Schema::Mbo, Compression::None, MboMsg::default())]
     #[case::uncompressed_trades_v1(1, Schema::Trades, Compression::None, TradeMsg::default())]
-    #[case::uncompressed_tbbo_v1(
-        1,
-        Schema::Tbbo,
-        Compression::None,
-        TbboMsg::default_for_schema(Schema::Tbbo)
-    )]
-    #[case::uncompressed_mbp1_v1(
-        1,
-        Schema::Mbp1,
-        Compression::None,
-        Mbp1Msg::default_for_schema(Schema::Tbbo)
-    )]
+    #[case::uncompressed_tbbo_v1(1, Schema::Tbbo, Compression::None, TbboMsg::default())]
+    #[case::uncompressed_mbp1_v1(1, Schema::Mbp1, Compression::None, Mbp1Msg::default())]
     #[case::uncompressed_mbp10_v1(1, Schema::Mbp10, Compression::None, Mbp10Msg::default())]
     #[case::uncompressed_ohlcv1d_v1(
         1,
@@ -793,18 +783,8 @@ mod tests {
     )]
     #[case::zstd_mbo_v1(1, Schema::Mbo, Compression::ZStd, MboMsg::default())]
     #[case::zstd_trades_v1(1, Schema::Trades, Compression::ZStd, TradeMsg::default())]
-    #[case::zstd_tbbo_v1(
-        1,
-        Schema::Tbbo,
-        Compression::ZStd,
-        TbboMsg::default_for_schema(Schema::Tbbo)
-    )]
-    #[case::zstd_mbp1_v1(
-        1,
-        Schema::Mbp1,
-        Compression::ZStd,
-        Mbp1Msg::default_for_schema(Schema::Mbp1)
-    )]
+    #[case::zstd_tbbo_v1(1, Schema::Tbbo, Compression::ZStd, TbboMsg::default())]
+    #[case::zstd_mbp1_v1(1, Schema::Mbp1, Compression::ZStd, Mbp1Msg::default())]
     #[case::zstd_mbp10_v1(1, Schema::Mbp10, Compression::ZStd, Mbp10Msg::default())]
     #[case::zstd_ohlcv1d_v1(
         1,
@@ -840,18 +820,8 @@ mod tests {
     #[case::zstd_statistics_v1(1, Schema::Statistics, Compression::ZStd, StatMsg::default())]
     #[case::uncompressed_mbo_v2(2, Schema::Mbo, Compression::None, MboMsg::default())]
     #[case::uncompressed_trades_v2(2, Schema::Trades, Compression::None, TradeMsg::default())]
-    #[case::uncompressed_tbbo_v2(
-        2,
-        Schema::Tbbo,
-        Compression::None,
-        TbboMsg::default_for_schema(Schema::Tbbo)
-    )]
-    #[case::uncompressed_mbp1_v2(
-        2,
-        Schema::Mbp1,
-        Compression::None,
-        Mbp1Msg::default_for_schema(Schema::Mbp1)
-    )]
+    #[case::uncompressed_tbbo_v2(2, Schema::Tbbo, Compression::None, TbboMsg::default())]
+    #[case::uncompressed_mbp1_v2(2, Schema::Mbp1, Compression::None, Mbp1Msg::default())]
     #[case::uncompressed_mbp10_v2(2, Schema::Mbp10, Compression::None, Mbp10Msg::default())]
     #[case::uncompressed_ohlcv1d_v2(
         2,
@@ -895,20 +865,11 @@ mod tests {
         Compression::None,
         StatMsg::default()
     )]
+    #[case::uncompressed_status_v2(2, Schema::Status, Compression::None, StatusMsg::default())]
     #[case::zstd_mbo_v2(2, Schema::Mbo, Compression::ZStd, MboMsg::default())]
     #[case::zstd_trades_v2(2, Schema::Trades, Compression::ZStd, TradeMsg::default())]
-    #[case::zstd_tbbo_v2(
-        2,
-        Schema::Tbbo,
-        Compression::ZStd,
-        TbboMsg::default_for_schema(Schema::Tbbo)
-    )]
-    #[case::zstd_mbp1_v2(
-        2,
-        Schema::Mbp1,
-        Compression::ZStd,
-        Mbp1Msg::default_for_schema(Schema::Mbp1)
-    )]
+    #[case::zstd_tbbo_v2(2, Schema::Tbbo, Compression::ZStd, TbboMsg::default())]
+    #[case::zstd_mbp1_v2(2, Schema::Mbp1, Compression::ZStd, Mbp1Msg::default())]
     #[case::zstd_cbbo_v2(
         2,
         Schema::Cbbo,
@@ -948,6 +909,7 @@ mod tests {
     )]
     #[case::zstd_imbalance_v2(2, Schema::Imbalance, Compression::ZStd, ImbalanceMsg::default())]
     #[case::zstd_statistics_v2(2, Schema::Statistics, Compression::ZStd, StatMsg::default())]
+    #[case::zstd_status_v2(2, Schema::Status, Compression::ZStd, StatusMsg::default())]
     fn test_dbn_identity<R: DbnEncodable + HasRType + PartialEq + Clone>(
         #[case] version: u8,
         #[case] schema: Schema,